pr: edit based on comments
