+++ server/my-strapi-project/.strapi-updater.json
@@ -1,5 +1,5 @@
 {
 	"latest": "5.0.1",
-	"lastUpdateCheck": 1727646235289,
+	"lastUpdateCheck": 1727820395442,
 	"lastNotification": 1727486950102
 }
